The whole project
This is the whole project, sorry about this. But I didn't use GitHub
since the start, so I have to commit all the files in once. This project
took me 41 days actually, because sometimes, some bugs couldn't be
solved in one day. But in each day's file, the OS improved a little,
from the very first day's boot function to the last day's whole system.
And also, I upload all the tools I used in this process. Mainly the qemu
simulater and some compiler, converter.

Hope you would like.